Modernize ASP.NET project
