+++ samples/Sample.AspNetCore/User.cs
@@ -1,0 +1,3 @@
+namespace Sample.AspNetCore;
+
+public record User(string Id, string Name, string Email);
