+++ samples/Sample.ConsoleApp/Services/UsersService.cs
@@ -11,7 +11,7 @@ public class UsersService
         _client = client;
     }
 
-    public async Task<User[]?> GetUsersAsync() => await _client.GetFromJsonAsync<User[]>("users");
+    public async Task<User[]?> GetUsersAsync(CancellationToken ct) => await _client.GetFromJsonAsync<User[]>("users", cancellationToken: ct);
 
     public record User(string Id, string Name, string Email);
 }
